make session.ts
