+++ Assemblies/ApiSites/Roblox.Auth.Api/Controllers/v1/login.ts
@@ -26,7 +26,7 @@
 */
 
 import { User } from '../../../../Platform/Membership/User'
-import { SessionUser } from '../../../../Platform/Membership/SessionUser'
+import { Session } from '../../../../Platform/Membership/Session'
 
 
 
@@ -39,7 +39,7 @@ export default {//TODO: FIX PARAMETERS
 		if (e == null) {
 			return res.status(403).send();
 		}
-		let ee = await SessionUser.CreateByID(<number>e.Id);
+		let ee = await Session.CreateByID(<number>e.Id);
 		return res.status(200).send(ee)
 	}
 };
